requirements.txt
behave==1.2.6
playwright==1.50.0
allure-behave==2.13.5
python-dotenv==1.0.1
pydantic==2.10.5
requests==2.32.3
